move PoolDataFetching into info_fetching.rs and rename to PoolInfoFetching
