+++ shared/src/balancer.rs
@@ -1,2 +1,3 @@
 pub mod event_handler;
+mod info_fetching;
 pub mod pool_fetching;
